+++ src/main/kotlin/com/github/warningimhack3r/npmupdatedependencies/backend/engine/NPMConfigReader.kt
@@ -99,8 +99,8 @@ class NPMConfigReader(project: Project) {
         companion object {
             private val log = logger<NPMConfigResolver>()
             private val npmHelpPathRegex = Regex("""^npm@[\d.]+ \S+$""")
-            private val scopeRegex = Regex("""^(@\S+):registry=(\S+)$""")
-            private val valueRegex = Regex("""^//(\S+):(\w+)=(\S+)$""")
+            private val scopeRegex = Regex("""^(@\S+):registry=(\S+)\s*$""")
+            private val valueRegex = Regex("""^//(\S+):(\w+)\s*=\s*(\S+)\s*$""")
             private val trailingSlashRegex = Regex("""/$""")
         }