+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Fixed
 
 - Fix shell commands hanging on Windows, regression from 3.7.0 (#186)
+- Fix crash with HTTP calls to registry without config info (#187)
 
 ## [3.7.0] - 2026-01-03