+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fix shell commands hanging on Windows, regression from 3.7.0 (#186)
 - Fix crash with HTTP calls to registry without config info (#187)
 - Avoid duplicate slashes in registry URLs
+- Ensure URLs parsed from configs are properly recognized when using additional whitespaces
 
 ## [3.7.0] - 2026-01-03