+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ## [Unreleased]
 
+## [3.7.1] - 2026-01-09
+
 ### Fixed
 
 - Fix shell commands hanging on Windows, regression from 3.7.0 (#186)
@@ -401,8 +403,9 @@
 
 - Initial release
 
-[Unreleased]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/3.7.0...HEAD
-[3.7.0]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/v3.6.1...3.7.0
+[Unreleased]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/v3.7.1...HEAD
+[3.7.1]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/v3.7.0...v3.7.1
+[3.7.0]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/v3.6.1...v3.7.0
 [3.6.1]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/v3.6.0...v3.6.1
 [3.6.0]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/v3.5.0...v3.6.0
 [3.5.0]: https://github.com/WarningImHack3r/npm-update-dependencies/compare/v3.4.0...v3.5.0